+++ src/publisher/pubMqtt.h
@@ -1,5 +1,5 @@
 //-----------------------------------------------------------------------------
-// 2023 Ahoy, https://ahoydtu.de
+// 2024 Ahoy, https://ahoydtu.de
 // Creative Commons - https://creativecommons.org/licenses/by-nc-sa/4.0/deed
 //-----------------------------------------------------------------------------
 
@@ -134,14 +134,14 @@ class PubMqtt {
             #endif
         }
 
-        bool tickerSun(uint32_t sunrise, uint32_t sunset, uint32_t offs) {
+        bool tickerSun(uint32_t sunrise, uint32_t sunset, uint16_t offsM, uint16_t offsE) {
             if (!mClient.connected())
                 return false;
 
             publish(subtopics[MQTT_SUNRISE], String(sunrise).c_str(), true);
             publish(subtopics[MQTT_SUNSET], String(sunset).c_str(), true);
-            publish(subtopics[MQTT_COMM_START], String(sunrise - offs).c_str(), true);
-            publish(subtopics[MQTT_COMM_STOP], String(sunset + offs).c_str(), true);
+            publish(subtopics[MQTT_COMM_START], String(sunrise - offsM).c_str(), true);
+            publish(subtopics[MQTT_COMM_STOP], String(sunset + offsE).c_str(), true);
 
             Inverter<> *iv;
             for(uint8_t i = 0; i < MAX_NUM_INVERTERS; i++) {
@@ -155,7 +155,7 @@ class PubMqtt {
 
 
             snprintf(mSubTopic, 32 + MAX_NAME_LENGTH, "comm_disabled");
-            publish(mSubTopic, (((*mUtcTimestamp > (sunset + offs)) || (*mUtcTimestamp < (sunrise - offs))) ? dict[STR_TRUE] : dict[STR_FALSE]), true);
+            publish(mSubTopic, (((*mUtcTimestamp > (sunset + offsE)) || (*mUtcTimestamp < (sunrise - offsM))) ? dict[STR_TRUE] : dict[STR_FALSE]), true);
 
             return true;
         }